mcp_server/pyproject.toml
[project]
name = "dart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "lxml>=5.3.2",
    "mcp[cli]==1.6",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]

[project.scripts]
mcp-server = "mcp_server:main"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
]
